[chat] added test that sends message to bob and alice
